+++ unique_sdk/unique_sdk/api_resources/_content.py
@@ -137,6 +137,7 @@ class Content(APIResource["Content"]):
         key: str
         title: Optional[str]
         mimeType: str
+        description: NotRequired[str | None]
         ownerType: NotRequired[str | None]
         ownerId: NotRequired[str | None]
         byteSize: NotRequired[int | None]
@@ -178,6 +179,7 @@ class Content(APIResource["Content"]):
         title: str | None
         metadata: Dict[str, Any] | None
         mimeType: str
+        description: str | None
         byteSize: int
         ownerId: str
         createdAt: str
@@ -349,6 +351,9 @@ class Content(APIResource["Content"]):
         """
         if "input" in params:
             params["input"]["metadata"] = params["input"].get("metadata") or {}
+            if "description" in params["input"] and not params["input"]["description"]:
+                params["input"].pop("description")
+
         return cast(
             "Content",
             cls._static_request(
@@ -372,6 +377,9 @@ class Content(APIResource["Content"]):
         """
         if "input" in params:
             params["input"]["metadata"] = params["input"].get("metadata") or {}
+            if "description" in params["input"] and not params["input"]["description"]:
+                params["input"].pop("description")
+
         return cast(
             "Content",
             await cls._static_request_async(